+++ analysis/functions.h
@@ -1,0 +1,8 @@
+#ifndef FUNCTIONS_H
+#define FUNCTIONS_H
+
+double norm(double value, double min, double max, double scale);
+double norm1(double value, double max);
+double norm100(double value, double max);
+
+#endif // FUNCTIONS_H