+++ drawerdp.h
@@ -30,11 +30,15 @@ protected:
     double secPitchDataDerivativeZero;
     int f0min, f0max;
     int userf0min, userf0max;
-    double rt, ru;
+    double range, userRange;
     double meanValueUMP, userMeanValueUMP;
+    double meanDerivativeValueUMP, userDerivativeMeanValueUMP;
     double centricGravityUMP, userCentricGravityUMP;
     double centricGravityUMP1, userCentricGravityUMP1;
     double centricGravityUMP2, userCentricGravityUMP2;
+    double centricGravityDerivativeUMP, userCentricGravityDerivativeUMP;
+    double centricGravityDerivativeUMP1, userCentricGravityDerivativeUMP1;
+    double centricGravityDerivativeUMP2, userCentricGravityDerivativeUMP2;
 
     SimpleGraphData * simple_data;
 };