+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.6] - 2019-02-11
+
+### Add
+- Save metrics for single file
+- Add new metrics
+
+### Change
+- Save relative metrics as 10 * log10
+
+## [0.4.5] - 2019-02-07
+
+### Fix
+- Fix gravity center UMP calculation
+
+### Add
+- Add gravity center UMP length metric
+
 ## [0.4.4] - 2019-02-05
 
 ### Change